Make it possible to iterate many times over a generator
This is the best solution I see to the problem with Java caching lambdas it
thinks are stateless.
